IAP: Term minutes (TEST) — a 10-minute subscription for renewal testing
Amazon's shortest store period is Weekly, but renewals are invisible in
receipts anyway — the SDK derives periods locally from the receipt's
PurchaseDate and the configured term, so a short TEST term makes a real
renewal observable on device within minutes, against a real App Tester
receipt and real elapsed time (no clock faking, the trusted-time
ratchet stays intact).

TestTermMinutes (> 0 replaces Term (days) in the period math) flows
settings -> catalog -> calculator: both setting models and both
converters, the settings window field with a warning box, and
IapConfiguredProduct.EffectiveTermDays expressing minutes as
fractional days. IapSubscriptionPeriods takes the term as double — the
math was double-based already, only signatures change. TermDays stays
mandatory (IAP-15) and grace still uses whole days.

Guard rails: the catalog logs a warning on every Configure while a
test term is active — it must be 0 in a production config or players
would receive periods every few minutes.

Panel: periods fire on reconcile, not on a timer, so a CHECK PERIODS
button runs one manually (otherwise the next period waits for a
background/foreground cycle); SIMULATE RENEWAL and the status line use
the effective term (status shows "10min (TEST)").

+++ Runtime/Modules/InAppPurchase/Core/IapSubscriptionPeriods.cs
@@ -10,6 +10,8 @@ namespace AMZNGoDSDK.Runtime
     /// (ReceiptId один на всю подписку, purchaseDate — дата ПЕРВОЙ покупки, cancelDate
     /// пуст). Поэтому периоды считаются: якорь — PurchaseDate чека, длина — TermDays из
     /// настроек, «сейчас» — доверенное время (SdkTrustedTime, не часы устройства).
+    /// Срок принимается в днях и может быть дробным: тестовый режим (TestTermMinutes)
+    /// выражает «подписку на 10 минут» как 10/1440 дня — математика и так на double.
     ///
     /// Граница периода N наступает в момент PurchaseDate + N × TermDays; период N считается
     /// оплаченным, если его граница уже наступила и лежит СТРОГО раньше потолков:
@@ -32,7 +34,7 @@ namespace AMZNGoDSDK.Runtime
         /// </summary>
         public static List<int> PeriodsToFire(
             long purchaseDateMs,
-            int termDays,
+            double termDays,
             long cancelDateMs,
             long deferredDateMs,
             DateTime trustedNowUtc,
@@ -63,7 +65,7 @@ namespace AMZNGoDSDK.Runtime
         /// <summary>Номер последнего наступившего оплаченного периода; -1, если ни одна граница не валидна.</summary>
         public static int CurrentPeriodIndex(
             long purchaseDateMs,
-            int termDays,
+            double termDays,
             long cancelDateMs,
             long deferredDateMs,
             DateTime trustedNowUtc)
@@ -78,15 +80,15 @@ namespace AMZNGoDSDK.Runtime
             return index;
         }
 
-        public static DateTime PeriodStartUtc(long purchaseDateMs, int termDays, int index) =>
-            FromUnixMs(purchaseDateMs).AddDays((double)termDays * index);
+        public static DateTime PeriodStartUtc(long purchaseDateMs, double termDays, int index) =>
+            FromUnixMs(purchaseDateMs).AddDays(termDays * index);
 
         /// <summary>
         /// Последний индекс, чья граница наступает СТРОГО раньше потолка. Граница ровно в
         /// момент потолка уже не оплачена: отменённая в точный конец периода подписка не
         /// получает следующий.
         /// </summary>
-        private static int LastIndexBeforeCeiling(DateTime purchaseUtc, int termDays, long ceilingMs)
+        private static int LastIndexBeforeCeiling(DateTime purchaseUtc, double termDays, long ceilingMs)
         {
             if (ceilingMs <= 0)
                 return int.MaxValue;   // потолка нет
